FrmPersonel den FrmProfil e kadar olan kısımda iş katmanı ile baplantılar ayarlandı.

+++ Market.Entity/Concrete/Iade.cs
@@ -7,7 +7,7 @@ namespace Market.Entity.Concrete
         public int Firma { get; set; }
         public int Bayi { get; set; }
         public int Fis { get; set; }
-        public int Satis { get; set; }
+        public long Satis { get; set; }
         public int Personel { get; set; }
         public DateTime TarihSaat { get; set; }
         public string Aciklama { get; set; }
